Updated README.md
Since the project is now customisable, the old README (which was just about GME and candlestick charts) needed to be updated. Well, it is now updated and has more info on the different features of the project, libraries used and some bonus info on how to read the different charts involved.